% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_plots.R
\name{clust_describe}
\alias{clust_describe}
\title{Describe clusters of an umapscan object}
\usage{
clust_describe(
  us,
  parent = "",
  labels = TRUE,
  type = c("boxplot", "ridges", "barplot", "keyness"),
  position = "fill",
  keyness_measure = c("chi2", "lr", "exact", "pmi"),
  n_terms = 20,
  text_size = 10,
  free_scale = TRUE,
  show_negative = TRUE
)
}
\arguments{
\item{us}{umapscan object to describe clusters}

\item{parent}{name of the parent cluster}

\item{labels}{if TRUE, use cluster label instead of id if available}

\item{type}{plot type, either \code{"boxplot"}, \code{"ridges"}, \code{"barplot"} or \code{"keyness"}}

\item{position}{if type = "barplot", \code{position} argument to add to \code{geom_bar}}

\item{keyness_measure}{if type = "keyness", passed as \code{measure} argument to \code{quanteda::textstat_keyness}}

\item{n_terms}{if type = "keyness", number of terms to show}

\item{text_size}{if type = "keyness", plot text size}

\item{free_scale}{if type = "keyness", don't use the same scale on each barplot}

\item{show_negative}{if TRUE and type = "keyness", show negative keyness features}
}
\description{
Describe clusters of an umapscan object
}
\details{
"boxplot" and "ridges" are suitable for continuous variables, "barplot" is better for
binary categorical ones. "keyness" is suitable when the original data is a document-feature
matrix.
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
us <- new_umapscan(iris_num, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- clust_compute(us, minPts = 3, eps = 0.5)
clust_describe(us)
us <- clust_compute(us, minPts = 3, eps = 0.45, parent = "3")
clust_describe(us, type = "ridge")
clust_describe(us, parent = "3")
}
\seealso{
\code{\link[=clust_compute]{clust_compute()}}, \code{\link[=clust_get_data]{clust_get_data()}}, \code{\link[=clust_members]{clust_members()}},
\code{\link[=clust_rename]{clust_rename()}}
}
