% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust_get_data}
\alias{clust_get_data}
\title{Get an umapscan cluster data}
\usage{
clust_get_data(us, id)
}
\arguments{
\item{us}{umapscan object}

\item{id}{cluster id to retrieve}
}
\value{
1 data frame with the \code{data} and \code{data_sup} variables of the cluster observations
}
\description{
Get an umapscan cluster data
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
iris_sup <- iris \%>\% select(Species)
us <- new_umapscan(iris_num, data_sup = iris_sup, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- clust_compute(us, minPts = 3, eps = 0.5)
clust_get_data(us, "1")
}
\seealso{
\code{\link[=clust_compute]{clust_compute()}}, \code{\link[=clust_describe]{clust_describe()}}, \code{\link[=clust_rename]{clust_rename()}}
}
