% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust_membership}
\alias{clust_membership}
\title{Get cluster membership for each observation of an umapscan object}
\usage{
clust_membership(
  us,
  parent = "",
  max_level,
  noise_inherit_parent = FALSE,
  labels = TRUE
)
}
\arguments{
\item{us}{umapscan object to describe clusters}

\item{parent}{name of the parent cluster}

\item{max_level}{get membership at most this level deep}

\item{noise_inherit_parent}{if TRUE, 'Noise' points are given their parent cluster
membership}

\item{labels}{if TRUE, use label when possible instead of identifier as cluster names}
}
\description{
The clusters taken into account are all the "leaves" of the object
clusters tree. If a \code{parent} is specified, only the leaves of its
subtree are taken.
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
us <- new_umapscan(iris_num, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- clust_compute(us, minPts = 3, eps = 0.5)
clust_membership(us)
us <- clust_compute(us, minPts = 3, eps = 0.45, parent = "3")
clust_membership(us)
clust_membership(us, parent = "3")

}
\seealso{
\code{\link[=clust_compute]{clust_compute()}}, \code{\link[=clust_get_data]{clust_get_data()}}, \code{\link[=clust_rename]{clust_rename()}}
}
