% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_plots.R
\name{clust_plot}
\alias{clust_plot}
\title{Plot clusters of an umapscan object}
\usage{
clust_plot(
  us,
  parent = "",
  noise_inherit_parent = FALSE,
  alpha = 1,
  ellipses = TRUE,
  fixed = FALSE,
  labels = TRUE
)
}
\arguments{
\item{us}{umapscan object}

\item{parent}{name of the parent cluster}

\item{noise_inherit_parent}{if TRUE, 'Noise' points are given their parent cluster
membership}

\item{alpha}{point transparency for clusters plot}

\item{ellipses}{if TRUE, plot confidence ellipses around clusters}

\item{fixed}{if TRUE, force coord_fixed on plot}

\item{labels}{if TRUE, use cluster label instead of id if available}
}
\description{
Plot clusters of an umapscan object
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
us <- new_umapscan(iris_num, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- clust_compute(us, minPts = 3, eps = 0.5, graph = FALSE)
clust_plot(us, alpha = 1)
us <- clust_compute(us, minPts = 3, eps = 0.45, graph = FALSE, parent = "3")
clust_plot(us, alpha = 0.5, ellipses = FALSE, parent = "3")
clust_plot(us)

}
\seealso{
\code{\link[=clust_compute]{clust_compute()}}, \code{\link[=clust_describe]{clust_describe()}}
}
