% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust_reinit}
\alias{clust_reinit}
\title{Reinit all clusters of an umapscan object}
\usage{
clust_reinit(us)
}
\arguments{
\item{us}{umapscan object}
}
\value{
An updated umapscan object with an empty clusters element.
}
\description{
Reinit all clusters of an umapscan object
}
