% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust_remove_noise}
\alias{clust_remove_noise}
\title{Remove a 'Noise' cluster from an umapscan object}
\usage{
clust_remove_noise(us, parent)
}
\arguments{
\item{us}{umapscan object}

\item{parent}{parent of the 'Noise' cluster to remove}
}
\description{
Remove a 'Noise' cluster from an umapscan object
}
