% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{collapse_clusters}
\alias{collapse_clusters}
\title{Collapse a clusters tree}
\usage{
collapse_clusters(us, collapsed)
}
\arguments{
\item{us}{an umapscan object}

\item{collapsed}{list of nodes to collapse}
}
\value{
An updated umapscan object with \code{collapsed} nodes children removed.
}
\description{
Collapse a clusters tree
}
