% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{compute_clusters}
\alias{compute_clusters}
\title{Compute DBSCAN clusters from a umapscan object}
\usage{
compute_clusters(us, eps, minPts, graph = TRUE, alpha = 0.5)
}
\arguments{
\item{us}{umapscan object}

\item{eps}{\code{eps} argument passed to \code{\link[dbscan:dbscan]{dbscan::dbscan()}}}

\item{minPts}{\code{minPts} argument passed to \code{\link[dbscan:dbscan]{dbscan::dbscan()}}}

\item{graph}{if TRUE, display a plot of the computed clusters}

\item{alpha}{point transparency for clusters plot}
}
\value{
Returns an updated \code{umapscan} object, and optionally displays a clusters plot.
}
\description{
This function runs DBSCAN with the specified arguments to compute new clusters
on a umapscan object. Only point which don't have an already validated cluster
are taken into account.
}
\examples{
iris_num <- iris \%>\% select_if(is.numeric)
iris_sup <- iris \%>\% select(Species)
us <- new_umapscan(iris_num, data_sup = iris_sup, n_neighbors = 25, min_dist = 0.1)
compute_clusters(us, minPts = 10, eps = 0.72, alpha = 1)

}
\seealso{
\code{\link[=new_umapscan]{new_umapscan()}}, \code{\link[=describe_clusters]{describe_clusters()}}, \code{\link[=get_cluster]{get_cluster()}}, \code{\link[=label_cluster]{label_cluster()}}
}
