% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{describe_clusters}
\alias{describe_clusters}
\title{Describe clusters of an umapscan object}
\usage{
describe_clusters(us, parent = "", type = c("boxplot", "ridges"))
}
\arguments{
\item{us}{umapscan object to describe clusters}

\item{parent}{name of the parent cluster}

\item{type}{plot type, either \code{"boxplot"} or \code{"ridges"}}
}
\description{
Describe clusters of an umapscan object
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
us <- new_umapscan(iris_num, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- compute_clusters(us, minPts = 3, eps = 0.5)
describe_clusters(us)
us <- compute_clusters(us, minPts = 3, eps = 0.45, parent = "3")
describe_clusters(us, type = "ridge")
describe_clusters(us, parent = "3")
}
\seealso{
\code{\link[=compute_clusters]{compute_clusters()}}, \code{\link[=get_cluster_data]{get_cluster_data()}}, \code{\link[=get_clusters_membership]{get_clusters_membership()}},
\code{\link[=rename_cluster]{rename_cluster()}}
}
