% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_cluster}
\alias{get_cluster}
\title{Get an umapscan cluster data}
\usage{
get_cluster(us, clust)
}
\arguments{
\item{us}{umapscan object}

\item{clust}{cluster to retrieve}
}
\value{
1 data frame with the \code{data} and \code{data_sup} variables of the cluster observations
}
\description{
Get an umapscan cluster data
}
\examples{
iris_num <- iris \%>\% select_if(is.numeric)
iris_sup <- iris \%>\% select(Species)
us <- new_umapscan(iris_num, data_sup = iris_sup, n_neighbors = 25, min_dist = 0.1)
us <- compute_clusters(us, minPts = 10, eps = 0.72, alpha = 1)
get_cluster(us, 1)
}
\seealso{
\code{\link[=compute_clusters]{compute_clusters()}}, \code{\link[=describe_clusters]{describe_clusters()}}, \code{\link[=label_cluster]{label_cluster()}}
}
