% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_cluster_data}
\alias{get_cluster_data}
\title{Get an umapscan cluster data}
\usage{
get_cluster_data(us, cluster)
}
\arguments{
\item{us}{umapscan object}

\item{cluster}{cluster to retrieve}
}
\value{
1 data frame with the \code{data} and \code{data_sup} variables of the cluster observations
}
\description{
Get an umapscan cluster data
}
\examples{
iris_num <- iris \%>\% select_if(is.numeric)
iris_sup <- iris \%>\% select(Species)
us <- new_umapscan(iris_num, data_sup = iris_sup, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- compute_clusters(us, minPts = 3, eps = 0.7)
get_cluster_data(us, "1")
}
\seealso{
\code{\link[=compute_clusters]{compute_clusters()}}, \code{\link[=describe_clusters]{describe_clusters()}}, \code{\link[=rename_cluster]{rename_cluster()}}
}
