% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_clusters_membership}
\alias{get_clusters_membership}
\title{Get cluster membership for each observation of an umapscan object}
\usage{
get_clusters_membership(
  us,
  parent = "",
  max_level,
  noise_inherit_parent = FALSE,
  labels = TRUE
)
}
\arguments{
\item{us}{umapscan object to describe clusters}

\item{parent}{name of the parent cluster}

\item{max_level}{get membership at most this level deep}

\item{noise_inherit_parent}{if TRUE, 'Noise' points are given their parent cluster
membership}

\item{if}{TRUE, use label when possible instead of identifier as cluster names}
}
\description{
The clusters taken into account are all the "leaves" of the object
clusters tree. If a \code{parent} is specified, only the leaves of its
subtree are taken.
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
us <- new_umapscan(iris_num, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- compute_clusters(us, minPts = 3, eps = 0.5)
get_clusters_membership(us)
us <- compute_clusters(us, minPts = 3, eps = 0.45, parent = "3")
get_clusters_membership(us)
get_clusters_membership(us, parent = "3")
}
\seealso{
\code{\link[=compute_clusters]{compute_clusters()}}, \code{\link[=get_cluster_data]{get_cluster_data()}}, \code{\link[=rename_cluster]{rename_cluster()}}
}
