% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_members}
\alias{get_members}
\title{Get members from a node name}
\usage{
get_members(us, node)
}
\arguments{
\item{us}{an umapscan object}

\item{node}{node to get members from}
}
\description{
Get members from a node name
}
