% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_noise_ids}
\alias{get_noise_ids}
\title{Get ids from a 'Noise' child of a node}
\usage{
get_noise_ids(us, node)
}
\arguments{
\item{us}{an umapscan object}

\item{node}{node to get child 'Noise' ids from}
}
\description{
Get ids from a 'Noise' child of a node
}
