% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_noise_members}
\alias{get_noise_members}
\title{Get members from a 'Noise' child of a node}
\usage{
get_noise_members(us, node)
}
\arguments{
\item{us}{an umapscan object}

\item{node}{node to get child 'Noise' members from}
}
\description{
Get members from a 'Noise' child of a node
}
