% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness.R
\name{keyness_stats}
\alias{keyness_stats}
\title{Generate cluster keyness statistics from a clustering result}
\usage{
keyness_stats(
  groups,
  data,
  keyness_measure = c("chi2", "lr", "exact", "pmi"),
  n_terms = 15,
  show_negative = TRUE,
  max_p = 0.05
)
}
\arguments{
\item{groups}{groups membership computed by \code{get_clusters_membership}}

\item{data}{object used to compute the clustering}

\item{keyness_measure}{statistics to compute}

\item{n_terms}{number of terms to display in keyness plots}

\item{show_negative}{if TRUE, show negative keyness features}

\item{max_p}{maximum keyness statistic p-value}
}
\description{
Generate cluster keyness statistics from a clustering result
}
\examples{
\dontrun{
library(quanteda)
corpus <- data_corpus_inaugural
dtm <- dfm(corpus, remove = stopwords("en"), tolower = TRUE, remove_punct = TRUE)
dtm <- dfm_trim(dtm, min_termfreq = 3)
res <- rainette(dtm, k = 3)
groups <- cutree_rainette(res, k = 3)
rainette_stats(groups, dtm)
}

}
\seealso{
\code{\link[quanteda:textstat_keyness]{quanteda::textstat_keyness()}}
}
