% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{label_cluster}
\alias{label_cluster}
\title{Label and validate an umapscan cluster}
\usage{
label_cluster(us, clust, label)
}
\arguments{
\item{us}{umapscan object}

\item{clust}{cluster identifier}

\item{label}{label to attribute to cluster}
}
\value{
A updated umapscan object
}
\description{
Label and validate an umapscan cluster
}
\examples{
iris_num <- iris \%>\% select_if(is.numeric)
iris_sup <- iris \%>\% select(Species)
us <- new_umapscan(iris_num, data_sup = iris_sup, n_neighbors = 25, min_dist = 0.1)
us <- compute_clusters(us, minPts = 10, eps = 0.72, alpha = 1)
us <- label_cluster(us, 1, "Cluster 1")
us
}
\seealso{
\code{\link[=compute_clusters]{compute_clusters()}}, \code{\link[=describe_clusters]{describe_clusters()}}, \code{\link[=get_cluster]{get_cluster()}}
}
