% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapscan.R
\name{plot.umapscan}
\alias{plot.umapscan}
\title{Plot an umapscan object}
\usage{
\method{plot}{umapscan}(
  x,
  color,
  palette,
  label = NULL,
  ellipses = FALSE,
  alpha = 0.2,
  fixed = FALSE,
  ...
)
}
\arguments{
\item{x}{umapscan object to be plotted}

\item{color}{variable to highlight points (\code{tidylang} style). Must be a variable
of the \code{data} or \code{data_sup} elements of \code{us}}

\item{palette}{\code{viridis} (for continuous \code{color}) or \code{brewer} (for categorical
\code{color}) palette name}

\item{label}{legend label}

\item{ellipses}{if TRUE and \code{color} is categorical, add repartition ellipses}

\item{alpha}{points transparency}

\item{fixed}{if TRUE, add \code{\link[ggplot2:coord_fixed]{ggplot2::coord_fixed()}} to plot}

\item{...}{arguments passed to other methods}
}
\description{
Plot an umapscan object
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
iris_sup <- iris \%>\% select(Species)
us <- new_umapscan(iris_num, data_sup = iris_sup, n_neighbors = 25, min_dist = 0.1)
plot(us, color = Species, ellipses = TRUE)
plot(us, color = Sepal.Width, palette = "viridis", alpha = 0.5, fixed = TRUE)

}
