% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_plots.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clusters of an umapscan object}
\usage{
plot_clusters(
  us,
  parent = "",
  noise_inherit_parent = FALSE,
  alpha = 1,
  ellipses = TRUE,
  fixed = FALSE
)
}
\arguments{
\item{us}{umapscan object}

\item{parent}{name of the parent cluster}

\item{noise_inherit_parent}{if TRUE, 'Noise' points are given their parent cluster
membership}

\item{alpha}{point transparency for clusters plot}

\item{ellipses}{if TRUE, plot confidence ellipses around clusters}

\item{fixed}{if TRUE, force coord_fixed on plot}
}
\description{
Plot clusters of an umapscan object
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
us <- new_umapscan(iris_num, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- compute_clusters(us, minPts = 3, eps = 0.5, graph = FALSE)
plot_clusters(us, alpha = 1)
us <- compute_clusters(us, minPts = 3, eps = 0.45, graph = FALSE, parent = "3")
plot_clusters(us, alpha = 0.5, ellipses = FALSE, parent = "3")
plot_clusters(us)

}
\seealso{
\code{\link[=compute_clusters]{compute_clusters()}}, \code{\link[=describe_clusters]{describe_clusters()}}
}
