% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapscan.R
\name{print.umapscan}
\alias{print.umapscan}
\title{Print an umapscan object}
\usage{
\method{print}{umapscan}(x, ...)
}
\arguments{
\item{x}{umapscan object to be printed}

\item{...}{arguments passed to other methods}
}
\description{
Print an umapscan object
}
