% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qualitative_palette}
\alias{qualitative_palette}
\title{Return a qualitative color palette depending on the number of values
of its \code{var} argument}
\usage{
qualitative_palette(var, label)
}
\description{
Return a qualitative color palette depending on the number of values
of its \code{var} argument
}
