% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{reinit_clusters}
\alias{reinit_clusters}
\title{Reinit all clusters of an umapscan object}
\usage{
reinit_clusters(us)
}
\arguments{
\item{us}{umapscan object}
}
\value{
An updated umapscan object with an empty clusters element.
}
\description{
Reinit all clusters of an umapscan object
}
