% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{remove_cluster}
\alias{remove_cluster}
\title{Remove a cluster from an umapscan object}
\usage{
remove_cluster(us, cluster)
}
\arguments{
\item{us}{umapscan object}

\item{cluster}{label of the cluster to remove}
}
\value{
An updated umapscan object (invisibly). Note that the original umapscan object is
modified in place.
}
\description{
If the cluster has children, they will be removed too.
}
\examples{
iris_num <- iris \%>\% select_if(is.numeric)
us <- new_umapscan(iris_num, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- compute_clusters(us, minPts = 3, eps = 0.7)
us <- compute_clusters(us, parent = "3" ,minPts = 3, eps = 0.45)
remove_cluster(us, "3_1")
us
remove_cluster(us, "3")
us
}
\seealso{
\code{\link[=compute_clusters]{compute_clusters()}}, \code{\link[=describe_clusters]{describe_clusters()}}
}
