% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{remove_noise_cluster}
\alias{remove_noise_cluster}
\title{Remove a 'Noise' cluster from an umapscan object}
\usage{
remove_noise_cluster(us, parent)
}
\arguments{
\item{us}{umapscan object}

\item{parent}{parent of the 'Noise' cluster to remove}
}
\description{
Remove a 'Noise' cluster from an umapscan object
}
