% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{rename_cluster}
\alias{rename_cluster}
\title{Rename an umapscan cluster identifier}
\usage{
rename_cluster(us, old, new)
}
\arguments{
\item{us}{umapscan object}

\item{old}{current cluster identifier}

\item{new}{new cluster identifier}
}
\value{
An updated umapscan object (invisibly). If two clusters have the same identifier after
renaming, they are merged together.
}
\description{
Rename an umapscan cluster identifier
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
us <- new_umapscan(iris_num, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- compute_clusters(us, minPts = 3, eps = 0.5)
rename_cluster(us, "1", "Cluster 1")
us
}
\seealso{
\code{\link[=label_cluster]{label_cluster()}}
}
