% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapscan_explor.R
\name{umapscan_explor}
\alias{umapscan_explor}
\title{Launch a shiny interface for umapscan object and clustering exploration}
\usage{
umapscan_explor(us)
}
\arguments{
\item{us}{an umapscan object}
}
\description{
Launch a shiny interface for umapscan object and clustering exploration
}
\examples{
\dontrun{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
iris_sup <- iris \%>\% select(Species)
us <- new_umapscan(iris_num, data_sup = iris_sup, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- clust_compute(us, minPts = 3, eps = 0.7)
us <- clust_compute(us, minPts = 3, eps = 0.45, parent = "3")
us <- clust_compute(us, minPts = 3, eps = 0.3, parent = "1")
umapscan_explor(us)
}

}
