% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explor_umapscan.R
\name{umapscan_tree}
\alias{umapscan_tree}
\title{Convert a clusters element of an umapscan object to a collapsibleTree}
\usage{
umapscan_tree(us, max_label_length = 100)
}
\arguments{
\item{us}{an umapscan object}

\item{max_label_length}{maximum length of tree node labels}
}
\description{
Convert a clusters element of an umapscan object to a collapsibleTree
}
