% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{describe_clusters}
\alias{describe_clusters}
\title{Describe current clusters of an umapscan object}
\usage{
describe_clusters(us, type = c("boxplot", "ridges"))
}
\arguments{
\item{us}{umapscan object to describe clusters}

\item{type}{plot type, either \code{"boxplot"} or \code{"ridges"}}
}
\description{
Only the not already validated clusters are taken into account.
}
\examples{
iris_num <- iris \%>\% select_if(is.numeric)
iris_sup <- iris \%>\% select(Species)
us <- new_umapscan(iris_num, data_sup = iris_sup, n_neighbors = 25, min_dist = 0.1)
us <- compute_clusters(us, minPts = 10, eps = 0.72, alpha = 1)
describe_clusters(us)
}
\seealso{
\code{\link[=compute_clusters]{compute_clusters()}}, \code{\link[=get_cluster]{get_cluster()}}, \code{\link[=label_cluster]{label_cluster()}}
}
