% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_members}
\alias{get_members}
\title{Get members from a node name}
\usage{
get_members(us, id)
}
\arguments{
\item{us}{an umapscan object}

\item{id}{cluster id to get members from}
}
\description{
Get members from a node name
}
