% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_noise_ids}
\alias{get_noise_ids}
\title{Get ids from a <Noise> child of a node
(not exported)}
\usage{
get_noise_ids(us, node)
}
\description{
Get ids from a <Noise> child of a node
(not exported)
}
