% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness.R
\name{keyness_stats}
\alias{keyness_stats}
\title{Generate cluster keyness statistics from a clustering result}
\usage{
keyness_stats(
  groups,
  data,
  keyness_measure = c("chi2", "lr", "exact", "pmi"),
  n_terms = 15,
  show_negative = TRUE,
  max_p = 0.05
)
}
\arguments{
\item{groups}{groups membership computed by \code{clust_membership}}

\item{data}{object used to compute the clustering}

\item{keyness_measure}{statistics to compute}

\item{n_terms}{number of terms to display in keyness plots}

\item{show_negative}{if TRUE, show negative keyness features}

\item{max_p}{maximum keyness statistic p-value}
}
\description{
Generate cluster keyness statistics from a clustering result
}
\seealso{
\code{\link[quanteda:textstat_keyness]{quanteda::textstat_keyness()}}
}
