% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{label_cluster}
\alias{label_cluster}
\title{Label an umapscan cluster}
\usage{
label_cluster(us, id, label)
}
\arguments{
\item{us}{umapscan object}

\item{id}{cluster identifier}

\item{label}{cluster label}
}
\value{
An updated umapscan object (invisibly).
}
\description{
Label an umapscan cluster
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
us <- new_umapscan(iris_num, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- compute_clusters(us, minPts = 3, eps = 0.5, graph = FALSE)
us <- rename_cluster(us, "1", "clust_1")
us <- label_cluster(us, "clust_1", "Cluster 1")
us$clusters
}
\seealso{
\code{\link[=rename_cluster]{rename_cluster()}}
}
