% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapscan.R
\name{new_umapscan}
\alias{new_umapscan}
\title{Compute UMAP embeddings from a numerical data frame}
\usage{
new_umapscan(
  d,
  n_neighbors = 15,
  min_dist = 0.01,
  metric = "euclidean",
  scale = FALSE,
  seed,
  data_sup = NULL,
  ...
)
}
\arguments{
\item{d}{a numerical matrix or data frame, or quanteda dfm}

\item{n_neighbors}{\code{n_neighbors} argument passed to \code{\link[uwot:umap]{uwot::umap()}}}

\item{min_dist}{\code{min_dist} argument passed to \code{\link[uwot:umap]{uwot::umap()}}}

\item{metric}{\code{metric} argument passed to \code{\link[uwot:umap]{uwot::umap()}}}

\item{scale}{scaling to be applied to the variables. Value passed to \code{\link[uwot:umap]{uwot::umap()}}.}

\item{seed}{integer seed passed to \code{\link[base:set.seed]{base::set.seed()}} before umap embeddings for
reproductibility}

\item{data_sup}{a data.frame of supplementary variables not used for UMAP
embeddings but for results visualisation}

\item{...}{other arguments passed to \code{\link[uwot:umap]{uwot::umap()}}}
}
\value{
A list of class \code{umapscan} with the following components :
\itemize{
\item \code{umap} : results of UMAP embeddings
\item \code{data} : original numerical dataset
\item \code{data_sup} : supplementary dataset
\item \code{clusters} : a tree-like hierarchy of clusters
\item \code{call} : function call
\item \code{seed} : the seed used for random numbers for all umap and dbscan operations
}
}
\description{
Compute UMAP embeddings from a numerical data frame
}
\examples{
library(dplyr)
iris_num <- iris \%>\% select_if(is.numeric)
iris_sup <- iris \%>\% select(Species)
new_umapscan(iris_num, data_sup = iris_sup)
}
\seealso{
\code{\link[uwot:umap]{uwot::umap()}}
}
