% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clusters of an umapscan object}
\usage{
plot_clusters(us, parent, alpha = 1, ellipses = TRUE)
}
\arguments{
\item{us}{umapscan object}

\item{parent}{name of the parent cluster}

\item{alpha}{point transparency for clusters plot}

\item{ellipses}{if TRUE, plot confidence ellipses around clusters}
}
\description{
Plot clusters of an umapscan object
}
\examples{
iris_num <- iris \%>\% select_if(is.numeric)
us <- new_umapscan(iris_num, n_neighbors = 25, min_dist = 0.1, seed = 1337)
us <- compute_clusters(us, minPts = 3, eps = 0.7, graph = FALSE)
plot_clusters(us, alpha = 1)
us <- compute_clusters(us, minPts = 3, eps = 0.45, graph = FALSE, parent = "3")
plot_clusters(us, alpha = 0.5, ellipses = FALSE, parent = "3")
plot_clusters(us)

}
\seealso{
\code{\link[=compute_clusters]{compute_clusters()}}, \code{\link[=describe_clusters]{describe_clusters()}}
}
