% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qualitative_palette}
\alias{qualitative_palette}
\title{Return a qualitative color palette depending on the number of values
of its \code{var} argument}
\usage{
qualitative_palette(var, label, type = c("color", "fill"))
}
\arguments{
\item{var}{vector to compute palette values from}

\item{label}{title of generated paletee}

\item{type}{type of color palette}
}
\description{
Return a qualitative color palette depending on the number of values
of its \code{var} argument
}
